+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+##### Changed
+* **\[core\]** Upgrade to react 16.2.0.
+
 ### [v0.20](https://github.com/react4j/react4j/tree/v0.20) (2017-12-01)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.19...v0.20)