+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   `props` and `context`.
 * **\[core\]** Remove `NativeComponent.setState(State)` method as `NativeComponent.setState(SetStateCallback)`
   is a better alternative.
+* **\[arez\]** Upgrade Arez to version 0.33.
 
 ### [v0.20](https://github.com/react4j/react4j/tree/v0.20) (2017-12-01)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.19...v0.20)